refactor: move helper to main module
